Bottom navigation bas added
