+++ .idea/misc.xml
@@ -13,12 +13,14 @@
         <entry key="..\:/Android_Apps/attil_Android-delivery/app/src/main/res/drawable/circle_normal_tick.xml" value="0.1365" />
         <entry key="..\:/Android_Apps/attil_Android-delivery/app/src/main/res/drawable/close_x.xml" value="0.1365" />
         <entry key="..\:/Android_Apps/attil_Android-delivery/app/src/main/res/drawable/custom_radio_drawable.xml" value="0.167" />
+        <entry key="..\:/Android_Apps/attil_Android-delivery/app/src/main/res/drawable/drawable_selector.xml" value="0.1885" />
         <entry key="..\:/Android_Apps/attil_Android-delivery/app/src/main/res/drawable/green_dot.xml" value="0.186" />
         <entry key="..\:/Android_Apps/attil_Android-delivery/app/src/main/res/drawable/ic_launcher_background.xml" value="0.2265" />
         <entry key="..\:/Android_Apps/attil_Android-delivery/app/src/main/res/drawable/ic_visibility_black_24dp.xml" value="0.1185" />
         <entry key="..\:/Android_Apps/attil_Android-delivery/app/src/main/res/drawable/ic_visibility_off_black_24dp.xml" value="0.187" />
         <entry key="..\:/Android_Apps/attil_Android-delivery/app/src/main/res/drawable/password_show_drawable.xml" value="0.1185" />
         <entry key="..\:/Android_Apps/attil_Android-delivery/app/src/main/res/layout/activity_home.xml" value="0.1421875" />
+        <entry key="..\:/Android_Apps/attil_Android-delivery/app/src/main/res/layout/activity_home_new.xml" value="0.176" />
         <entry key="..\:/Android_Apps/attil_Android-delivery/app/src/main/res/layout/activity_login.xml" value="0.264" />
         <entry key="..\:/Android_Apps/attil_Android-delivery/app/src/main/res/layout/activity_main.xml" value="0.19322916666666667" />
         <entry key="..\:/Android_Apps/attil_Android-delivery/app/src/main/res/layout/activity_splash.xml" value="0.22826086956521738" />
@@ -38,6 +40,7 @@
         <entry key="..\:/Android_Apps/attil_Android-delivery/app/src/main/res/layout/notification_list_adapter.xml" value="0.264" />
         <entry key="..\:/Android_Apps/attil_Android-delivery/app/src/main/res/layout/progress_bar_layout.xml" value="0.19322916666666667" />
         <entry key="..\:/Android_Apps/attil_Android-delivery/app/src/main/res/layout/view_order.xml" value="0.22135416666666666" />
+        <entry key="..\:/Android_Apps/attil_Android-delivery/app/src/main/res/menu/bottom_nav.xml" value="0.24895833333333334" />
         <entry key="..\:/Android_Apps/attil_Android-delivery/app/src/main/res/mipmap-anydpi-v26/ic_launcher.xml" value="0.2265" />
         <entry key="..\:/Android_Apps/attil_Android-delivery/app/src/main/res/mipmap-xhdpi/ic_launcher.xml" value="0.2265" />
         <entry key="..\:/Android_Apps/attil_Android-delivery/app/src/main/res/mipmap-xhdpi/ic_launcher_round.xml" value="0.167" />
