+++ app/src/main/res/color/color_selector.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:color="@color/purple_500" android:state_checked="true"/>
+    <item android:color="@color/grey" />
+</selector>
