+++ app/src/main/res/drawable/drawable_selector.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:drawable="@color/white" android:state_checked="true" />
+    <item android:drawable="@color/white" />
+</selector>
