+++ app/src/main/res/menu/bottom_nav.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item android:title="Home"
+        android:id="@+id/home"
+        android:icon="@drawable/compass"/>
+    <item android:title="Delivery"
+        android:id="@+id/delivery"
+        android:icon="@drawable/boxes"/>
+    <item android:title="Notification"
+        android:id="@+id/notity"
+        android:icon="@drawable/bell"/>
+    <item android:title="Profile"
+        android:id="@+id/profile"
+        android:icon="@drawable/profile"/>
+</menu>
